Create CTF Wall
[quest#231] CTF Wall +500

"Make a pull request to the fhq-server repository In the dev branch.

You are invited to leave a mark on the history of freehackquest. Add your message to the ctf-wall page for other users, indicating from whom the nickname and date. The flag will either be emailed to you or attached to a pull request. Be polite."